# contracts/openapi-v1-generated.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "https://virtserver.swaggerhub.com/guibirow/Yumido/v1"
    }
  ],
  "info": {
    "description": "This is an API to manage Reward in the restaurants",
    "version": "v1",
    "title": "Reward API",
    "contact": {
      "email": "stacks@amido.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/reward/{id}/category/{categoryId}": {
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "Update a category in the reward",
        "operationId": "updateRewardCategory",
        "description": "Update a category to reward",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "reward id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "Id for Category being removed",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "OpenID": [
              "admin"
            ]
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/Created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/ConflictItemExists"
          }
        },
        "requestBody": {
          "description": "Category being added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateCategory"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Removes a category and its items from reward",
        "operationId": "deleteCategory",
        "description": "Removes a category and its items from reward",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "reward id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "Id for Category being removed",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "OpenID": [
              "admin"
            ]
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrUpdateCategory": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Burguers"
          },
          "description": {
            "type": "string",
            "example": "A delicious selection of burguers"
          }
        }
      },
...
    },
    "responses": {
...
      "Created": {
        "description": "Resource created",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found"
      },
      "NoContent": {
        "description": "No Content"
      },
      "BadRequest": {
        "description": "Bad Request"
      },
      "Unauthorized": {
        "description": "Unauthorized, Access token is missing or invalid"
      },
      "Forbidden": {
        "description": "Forbidden, the user does not have permission to execute this operation"
      },
      "ConflictItemExists": {
        "description": "Conflict, an item already exists"
      }
    },
    "securitySchemes": {
      "OpenID": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://example.com/.well-known/openid-configuration"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}